Refactor to split up `workflow.py` into supporting files
